feat(useTernaryDarkMode): Add `defaultValue` parameter (#344)
* feat(useTernaryDarkMode): Add `defaultValue` parameter

This parameter will allow the user to set a default value for the dark mode when there's no value in the local storage. The function previously always defaulted to 'system'. With this change, the developer can now specify a different default value if desired. Documentation is also updated to reflect these changes.

* ci: add changeset for ternary dark mode hook default value change

* ✨ Update `useDarkMode` and `useTernaryDarkMode` signatures using function overload for smooth migration

- and update unit tests

---------

Co-authored-by: Julien <juliencaron@protonmail.com>